+++ migrations/000125_payment_settlement_results.down.sql
@@ -1,0 +1,2 @@
+DROP TABLE IF EXISTS payment_settlement_effects;
+DROP TABLE IF EXISTS payment_settlement_results;